// src-deprec/check/src/common/types/commerce.types.ts
export type PaginatedResult<T> = {
  data: T[]
  total: number
  page: number
  limit: number
  total_pages: number
  has_next: boolean
  has_prev: boolean
}